[[feature-stackarr]] MCObject: Add static initialize/finalize methods.
For stack array support, there are a number of important
class-specific create-once values that are required (in particular
MCTypeInfoRefs for the ExportState()/ImportState() API).

Add a general static initialization and finalization methods to
MCObject and each of its derived classes, and call them from X_open().

As part of this, change MCObject::GetStateTypeInfo() to store its
create-once value in a static class member.