libscript: Add MCScriptCreatePackageWithModules().
Add a new private function that takes a group of modules and collects
them into a package.  This works by creating a package, and then
making each module's package pointer to point to the package.  It
doesn't actually return the newly created package; this can be
obtained by calling MCScriptGetPackageOfModule().